Fix Type Errors In DBItems Component In Dashboard Screen

+++ client/src/components/DashboardComponents/DBItems.tsx
@@ -5,6 +5,7 @@ import { HiCurrencyDollar } from "../../assets/icons";
 import { setAllProducts } from "../../store";
 import { toast } from "react-hot-toast";
 import { deleteProduct, getAllProducts } from "../../api";
+import { Product } from "../../Types";
 
 const DBItems: React.FC = () => {
   const products = useAppSelector(({ generalSlice }) => generalSlice.products);
@@ -19,7 +20,7 @@ const DBItems: React.FC = () => {
           {
             title: "Image",
             field: "imageURL",
-            render: (rowData) => (
+            render: (rowData: Product) => (
               <img
                 src={rowData.imageURL}
                 className="w-32 h-16 object-contain rounded-md"
@@ -37,10 +38,10 @@ const DBItems: React.FC = () => {
           {
             title: "Price",
             field: "product_price",
-            render: (rowData) => (
+            render: (rowData: Product) => (
               <p className="text-xl font-semibold text-textColor flex items-center justify-center ">
                 <HiCurrencyDollar className="text-red-400" />
-                {parseFloat(rowData.product_price).toFixed(2)}
+                {rowData.product_price.toFixed(2)}
               </p>
             ),
           },
@@ -51,23 +52,25 @@ const DBItems: React.FC = () => {
           {
             icon: "edit",
             tooltip: "Edit Data",
-            onClick: (event, rowData) => {
+            onClick: (event: any, rowData: Product) => {
               alert("You want to edit " + rowData.product_name);
             },
           },
           {
             icon: "delete",
             tooltip: "Delete Data",
-            onClick: async (event, rowData) => {
+            onClick: async (event: any, rowData: Product) => {
               if (
                 window.confirm("Are you sure, you want to perform this action")
               ) {
                 try {
-                  await deleteProduct(rowData.productId);
+                  if(rowData.productId){
+                    await deleteProduct(rowData.productId);
                   toast.success("Product Deleted");
 
                   const allProducts = await getAllProducts();
                   dispatch(setAllProducts(allProducts));
+                  }
                 } catch (error: any) {
                   toast.error("Failed to delete the product: " + error.message);
                 }
